feat: Rename Packages to use io.meeds - Meeds-io/MIPs#165
