+++ component/web/src/main/java/io/meeds/social/handler/SpaceAccessHandler.java
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.exoplatform.social.core.space.impl;
+package io.meeds.social.handler;
 
 import static org.exoplatform.portal.application.PortalRequestHandler.REQUEST_SITE_NAME;
 import static org.exoplatform.portal.application.PortalRequestHandler.REQUEST_SITE_TYPE;
