+++ component/web/src/main/java/io/meeds/social/handler/SpacePermanentLinkHandler.java
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package io.meeds.social.portlet;
+package io.meeds.social.handler;
 
 import static io.meeds.social.permlink.plugin.SpacePermanentLinkPlugin.APPLICATION_URI;
 import static io.meeds.social.permlink.plugin.SpacePermanentLinkPlugin.OBJECT_TYPE;
