+++ webapp/pom.xml
@@ -161,17 +161,6 @@
 
     <dependency>
       <groupId>io.meeds.portal</groupId>
-      <artifactId>portal.component.web.api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.meeds.portal</groupId>
-      <artifactId>portal.component.web.api</artifactId>
-      <type>test-jar</type>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>io.meeds.portal</groupId>
       <artifactId>portal.component.web.resources</artifactId>
     </dependency>
     <dependency>
